TINDER-81: Fix typo in Roster
The ‘toString’ of the Roster class will include the word subscrption which arguably should be subscription.

toString wouldn’t typically be used for anything other than make humans understand the composition of the instance. In cases where this is not the case, users would benefit from having a correction (although existing users will see a breakage because of this change).