TINDER-82: Ignore empty formfield values
`org.xmpp.forms.DataForm#addField(java.lang.String, java.lang.String, org.xmpp.forms.FormField.Type)`'s implementation defines that the first argument (`variable`) is optional. It then checks using this method to see if it is set:

```
if (variable != null && variable.trim().length() >= 0) {
    result.setVariable(variable);
}
```

A similar check is done for the second argument (`label`).

The second condition in the if-statement is redundant: if the value is not null, it is assured to have at least 0 length.

This is probably an oversight: I assume that the intention is to check for ‘not empty’.

With the existing implementation, an empty value will cause the value to be used, which might lead to empty attributes or elements. I presume that those are undesirable. A “set to nothing” declaration would be far-fetched and equally defined by not having the attribute/label at all.